uhd: Align numeric data types in digital filter classes
Modified accessors and constructors in digital filter classes
to align with the type declared in the corresponding fields.

+++ host/lib/types/filters_python.hpp
@@ -12,6 +12,7 @@
 
 void export_filters(py::module& m)
 {
+
     using filter_info_base        = uhd::filter_info_base;
     using filter_info_type        = filter_info_base::filter_type;
     using analog_filter_base      = uhd::analog_filter_base;
@@ -56,10 +57,10 @@ void export_filters(py::module& m)
             bool,
             size_t,
             double,
-            size_t,
-            size_t,
-            double,
-            size_t,
+            uint32_t,
+            uint32_t,
+            int16_t,
+            uint32_t,
             std::vector<int16_t>>())
 
         // Methods
@@ -77,10 +78,10 @@ void export_filters(py::module& m)
             bool,
             size_t,
             double,
-            size_t,
-            size_t,
-            size_t,
-            size_t,
+            uint32_t,
+            uint32_t,
+            int16_t,
+            uint32_t,
             std::vector<int16_t>>())
 
         // Methods
